actionGrammar: address dispatch optimizer review feedback
- Fix DispatchPart docstring to accurately describe filter-only semantics
  (peeked token is not consumed; rules retain their full original parts).
- Share wordBoundaryScriptRe between matcher and optimizer (single source
  of truth) instead of duplicating the script list regex.
- Remove unused tokenEnd / originalIndex fields from DispatchBacktrack.
- Extract referencesPrefixBoundCanonical helper from checkFactoringEligible
  for clarity in the tail-disabled path.
- Drop redundant inner debug-log guard in dispatchifyAlternations.
- Lift printAligned / visibleLen ANSI-aware helpers into benchUtil so
  dfaBenchmark and other bench scripts can share them.
- Note in dfaBenchmark that missing grammar files are skipped silently.
- Rename misleading dispatch-spec test to describe what it actually
  asserts.
